Split models out into their own files in the models/ folder per request
